Rename Android to match iOS
