+++ Example/index.ios.js
@@ -12,21 +12,12 @@ var {
   View,
 } = React;
 
+var SignatureCapture = require('react-native-signature-capture');
+
 var Example = React.createClass({
   render: function() {
     return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>
-          Welcome to React Native!
-        </Text>
-        <Text style={styles.instructions}>
-          To get started, edit index.ios.js
-        </Text>
-        <Text style={styles.instructions}>
-          Press Cmd+R to reload,{'\n'}
-          Cmd+D or shake for dev menu
-        </Text>
-      </View>
+      <SignatureCapture />
     );
   }
 });
